Shutdown servers on shutdown signal
When closing a listener, the server won't accept any new connections.
However, we should still allow for the current connections / requests to
finish. Calling ``server.Shutdown`` allows us to wait for them to
finish. We'll be doing this only if drainConnectionsTimeout has been set
for the server.

Note that in the case we fail to bootstrap / join the cluster, we'll be
resetting a few things, including the cluster membership. This includes
the HTTPS and unix socket servers we have open by closing them.

However, we cannot gracefully shutdown the servers, as there's at least
one connection that is still open: the bootstrap / join request. Forcing
the connection to close before we're able to write the request response
will result in the client getting an EOF error, and no information
regarding the failure.

Running the revert actions in a goroutine will address this issue: while
the revert happens, we'll be able to return and write the HTTP response
and then close the connection, finally allowing the Servers to gracefully
shutdown, and the clients to be happy. As a bonus, this will also
significantly reduce the amount of time spent by the bootstrap / join
requests in case of failures.

Signed-off-by: Claudiu Belu <claudiu.belu@canonical.com>