internal/rest/resources: Return joinInfo even on error
When joining a cluster, if any error occurs after becoming a cluster
member, we're not removing the current node from the cluster, even
though we should.

That is because in joinWithToken after calling internalClient.AddClusterMember
and getting a joinInfo object, we may still return nil afterwards in
case of any error. Because of this, the controlPost reverter won't have
any joinInfo object to use for removing itself from the cluster.

Because of this issue, if the cluster join operations fails, that node
may no longer be able to join the cluster, as it appears to have already
joined the cluster.

This commit addresses this issue by returning the joinInfo object after
it exists.

Signed-off-by: Claudiu Belu <claudiu.belu@canonical.com>